requirements.txt
google-generativeai==0.3.0
flask==3.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
olefile==0.46
openpyxl==3.1.2
python-pptx==0.6.23
gunicorn==21.2.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
easyocr==1.7.0
Pillow==10.1.0
requests==2.31.0
notion-client==2.2.1
pymysql==1.1.0
